Update transpilation with latest C2Rust

+++ src/internal.rs
@@ -1,2 +1,2 @@
-pub const __CHAR_BIT__: libc::c_int = 8i32;
-pub const __INT_MAX__: libc::c_int = 2147483647i32;
+pub const __CHAR_BIT__: libc::c_int = 8 as libc::c_int;
+pub const __INT_MAX__: libc::c_int = 2147483647 as libc::c_int;
