+++ src/jconfig_h.rs
@@ -1,5 +1,5 @@
-pub const BITS_IN_JSAMPLE: libc::c_int = 8i32;
+pub const BITS_IN_JSAMPLE: libc::c_int = 8 as libc::c_int;
 /* Version ID for the JPEG library.
  * Might be useful for tests like "#if JPEG_LIB_VERSION >= 60".
  */
-pub const JPEG_LIB_VERSION: libc::c_int = 62i32;
+pub const JPEG_LIB_VERSION: libc::c_int = 62 as libc::c_int;
