+++ src/limits_h.rs
@@ -1,4 +1,3 @@
 pub const INT_MAX: libc::c_int = crate::internal::__INT_MAX__;
-pub const INT_MIN: libc::c_int = -crate::internal::__INT_MAX__ - 1i32;
-use crate::internal::__INT_MAX__;
-pub const CHAR_BIT: libc::c_int = 8i32;
+pub const INT_MIN: libc::c_int = -crate::internal::__INT_MAX__ - 1 as libc::c_int;
+pub const CHAR_BIT: libc::c_int = 8 as libc::c_int;
